+++ deltafi-common/src/main/java/org/deltafi/common/types/DeltaFile.java
@@ -19,6 +19,7 @@ package org.deltafi.common.types;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.*;
+import org.apache.commons.lang3.tuple.Pair;
 import org.deltafi.common.content.Segment;
 import org.deltafi.common.converters.KeyValueConverter;
 import org.jetbrains.annotations.NotNull;
@@ -46,6 +47,7 @@ public class DeltaFile {
   private List<String> childDids;
   private int requeueCount;
   private Long ingressBytes;
+  private Long referencedBytes;
   private Long totalBytes;
   private DeltaFileStage stage;
   private List<Action> actions;
@@ -366,6 +368,18 @@ public class DeltaFile {
     return builder.build();
   }
 
+  public List<Segment> referencedSegments() {
+    List<Segment> segments = getProtocolStack().stream()
+            .flatMap(p -> p.getContent().stream()).map(Content::getContentReference)
+            .flatMap(c -> c.getSegments().stream())
+            .collect(Collectors.toList());
+    segments.addAll(getFormattedData().stream()
+            .map(FormattedData::getContentReference)
+            .flatMap(f -> f.getSegments().stream())
+            .collect(Collectors.toList()));
+    return segments;
+  }
+
   public List<Segment> storedSegments() {
     List<Segment> segments = getProtocolStack().stream()
             .flatMap(p -> p.getContent().stream()).map(Content::getContentReference)
@@ -378,6 +392,27 @@ public class DeltaFile {
             .filter(s -> s.getDid().equals(getDid()))
             .collect(Collectors.toList()));
     return segments;
+  }
+
+  public void recalculateBytes() {
+    setReferencedBytes(recalculateBytes(referencedSegments()));
+    setTotalBytes(recalculateBytes(storedSegments()));
+  }
+
+  private static long recalculateBytes(List<Segment> inputSegments) {
+    // keep track of the first and last offset for each unique uuid contentReference
+    // make an assumption that we won't have disjoint segments
+    Map<String, Pair<Long, Long>> segments = new HashMap<>();
+    inputSegments.forEach(storedSegment -> {
+      if (segments.containsKey(storedSegment.getUuid())) {
+        Pair<Long, Long> segment = segments.get(storedSegment.getUuid());
+        segments.put(storedSegment.getUuid(), Pair.of(Math.min(segment.getLeft(), storedSegment.getOffset()), Math.max(segment.getRight(), storedSegment.getOffset() + storedSegment.getSize())));
+      } else {
+        segments.put(storedSegment.getUuid(), Pair.of(storedSegment.getOffset(), storedSegment.getOffset() + storedSegment.getSize()));
+      }
+    });
+
+    return segments.values().stream().map(p -> p.getRight() - p.getLeft()).mapToLong(Long::longValue).sum();
   }
 
   public void cancelQueuedActions() {