+++ migrations/migrate-referenced-bytes.js
@@ -1,0 +1,3 @@
+db.deltaFile.updateMany({referencedBytes: {$exists: false}}, [
+    {"$set": {"referencedBytes": "$totalBytes"}}
+])